matchmaker/requirements.txt
mysql-connector-python>=8.0.0
python-valve>=0.2.0
docker>=7.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
pytest-cov>=4.1.0
